# server/docs/swagger.yaml
openapi: 3.0.0
info:
  title: ByteStash API
  version: "1.0.0"
  description: API documentation for ByteStash snippet storage service.
servers:
  - url: /
    description: Local server
components:
  securitySchemes:
    JwtAuth:
      type: apiKey
      in: header
      name: bytestashauth
      description: JWT token obtained from /api/auth/login. Must be prefixed with "bearer TOKEN"
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for CLI access
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        ByteStash API key sent as a bearer token (Authorization: Bearer <key>).
        Used by the MCP endpoint, which most remote MCP clients authenticate with.
...
paths:
  /api/auth/login:
    post:
      summary: Authenticate a user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
                      created_at:
                        type: string
                      is_admin:
                        type: boolean
                      is_active:
                        type: integer
                        nullable: true
                      email:
                        type: string
                        nullable: true
                      name:
                        type: string
                        nullable: true
                      oidc_id:
                        type: string
                        nullable: true
                      oidc_provider:
                        type: string
                        nullable: true
        '401':
          description: Invalid credentials
  /api/auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successfully registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
                      created_at:
                        type: string
                      is_admin:
                        type: boolean
        '400':
          description: Validation error
...
  /api/auth/verify:
    get:
      summary: Verify a JWT token
      tags:
        - Authentication
      security:
        - JwtAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
                      created_at:
                        type: string
                      is_admin:
                        type: boolean
        '401':
          description: Invalid token
  /api/auth/anonymous:
    post:
      summary: Create an anonymous session
      tags:
        - Authentication
      responses:
        '200':
          description: Anonymous session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
                      created_at:
                        type: string
        '403':
          description: Anonymous login not allowed
...